Add ckip segmentation method

+++ data/segment_words.py
@@ -2,48 +2,70 @@ import argparse
 from glob import glob
 import os
 import pickle
+import sys
+
+sys.path.append('.')
 
 import jieba
 from tqdm import tqdm
 
+from ckipclient import CKIPClient
 from iscr.utils import load_from_pickle, save_to_pickle
 
-def segment_jieba(text):
-    return ' '.join(list(jieba.cut(text)))
-
-def segment_ckip(text):
-    raise NotImplementedError
+client = None  # For CKIP
 
 
-def segment_file(segment, filepath):
-    ret = ""
-    with open(filepath, 'r') as fin:
-        for line in fin.readlines():
-            line = ''.join(line.split()) + '\n'
-            ret += segment(line)
-    return ret
+def segment_jieba(text, verbose=False):
+	result = list(jieba.cut(text))
+	if verbose:
+		print(result)
+	return ' '.join(result)
+
+
+def segment_ckip(text, verbose=False):
+	result = client.segment(text, pos=False)
+	if verbose:
+		print(result[0])
+	return '\n'.join(result[0])
+
+
+def segment_file(segment, filepath, verbose=False):
+	text = ""
+	with open(filepath, 'r') as fin:
+		for line in fin.readlines():
+			text += ''.join(line.split()) + '\n'
+	ret = segment(text)
+	return ret
 
 
 def run_segment(segment, query_file, out_query_file, transcript_dir, out_transcript_dir):
     # Segment query
-    segmented_query = segment_file(segment, query_file)
-    with open(out_query_file, 'w') as fout:
-        fout.write(segmented_query)
+	print("Segmenting query...")
+	if os.path.exists(out_query_file):
+		print("{} already exists, skipping...".format(out_query_file))
+	else:
+		segmented_query = segment_file(segment, query_file, verbose=False)
+		with open(out_query_file, 'w') as fout:
+			fout.write(segmented_query)
 
     # Segment Documents
-    if not os.path.isdir(out_transcript_dir):
-        os.makedirs(out_transcript_dir)
+	if not os.path.isdir(out_transcript_dir):
+		os.makedirs(out_transcript_dir)
 
-    for filepath in tqdm(glob(os.path.join(transcript_dir, '*'))):
-        filename = os.path.basename(filepath)
-        segmented_transcript = segment_file(segment, filepath)
-        out_filename = os.path.join(out_transcript_dir, filename)
-        with open(out_filename, 'w') as fout:
-            fout.write(segmented_transcript)
+	print("Segmenting documents...")
+	for filepath in tqdm(glob(os.path.join(transcript_dir, '*'))):
+		filename = os.path.basename(filepath)
+		out_filename = os.path.join(out_transcript_dir, filename)
+		if os.path.exists(out_filename):
+			print("Skipping {}...".format(out_filename))
+			continue
+
+		segmented_transcript = segment_file(segment, filepath)
+		with open(out_filename, 'w') as fout:
+			fout.write(segmented_transcript)
 
 
 segment_dict = {
-    'unigram': segment_unigram,
     'jieba': segment_jieba,
     'ckip': segment_ckip,
 }
@@ -60,15 +82,12 @@ if __name__ == "__main__":
                         default='./data/query/PTV.utf8.jieba.query', help='segmented query location')
     parser.add_argument('--out_transcript_dir', type=str,
                         default='./data/PTV.debug.jieba', help='segmented query location')
-    parser.add_argument('-u', '--user', type=str,
-                        default='./data/query/query_dict_for_jieba.txt')
     args = parser.parse_args()
 
-    if args.user:
-        print("Loading {} for jieba...".format(args.user))
-        jieba.load_userdict(args.user)
-
     segment = segment_dict[args.segment]
+    if args.segment == 'ckip':
+    	client = CKIPClient('140.109.19.104', 1501,
+                        	'iammrhelo', 'hl4su3a8')
 
     run_segment(segment, args.query_file, args.out_query_file,
                 args.transcript_dir, args.out_transcript_dir)
